refactor: move merchant routes into dashboard shell
