+++ src/app/(dashboard)/invoice-tools/page.tsx
@@ -1,0 +1,23 @@
+import { InvoiceToolsClient } from "./invoice-tools-client";
+
+export const metadata = {
+  title: "Invoice tools | Shade",
+  description: "Validation, lookup, timeline, and avatar components preview.",
+};
+
+export default function InvoiceToolsPage() {
+  return (
+    <div className="mx-auto w-full max-w-5xl space-y-10">
+      <header className="space-y-1">
+        <p className="text-sm font-semibold uppercase text-primary">
+          Shade merchant
+        </p>
+        <h1 className="text-3xl font-bold">Invoice tools</h1>
+        <p className="text-sm text-muted-foreground">
+          Validated invoice creation, payer lookup, timeline, and logo uploader.
+        </p>
+      </header>
+      <InvoiceToolsClient />
+    </div>
+  );
+}
