+++ src/app/(dashboard)/settings/profile/page.tsx
